requirements.txt
langchain~=0.3.0
langchain-openai~=0.2.0
openai~=1.45.0
uvicorn~=0.30.6
fastapi~=0.114.2
pydantic~=2.9.1
streamlit~=1.38.0
requests~=2.32.3
python-multipart~=0.0.9
python-docx~=1.1.2
pandas~=2.2.2
